requirements.txt
Django==6.0.3
djangorestframework==3.15.2
celery==5.4.0
redis==5.0.8
psycopg2-binary==2.9.9
boto3==1.35.0
openai==0.28
django-storages==1.14.4
PyPDF2==3.0.1
dj-database-url==2.2.0
python-dotenv==1.0.1
django-cors-headers==4.0.0
requests==2.31.0
httpx